Add number below thumb of sliding scale, move sliding scale to separate folder.
