+++ frontend/political_scale/src/popUp/PopUp.component.js
@@ -1,13 +1,14 @@
 import React from 'react'
-import { PopupContainer,PopupText } from './PopUp.styled';
+import { PopupContainer } from './PopUp.styled';
+import SlidingScale from '../slidingScale/SlidingScale.component';
 
-const PopUp = ({isArticle}) => {
+const PopUp = ({isArticle, category, value}) => {
     //If user not on article then no pop up
     if(!isArticle) return null;
 
     return(
         <PopupContainer>
-            <PopupText>This is an article.fr.</PopupText>
+            <SlidingScale category={category} value = {value} />
         </PopupContainer>
     );
 };
